+++ src/render/lighting/lighting_norm.rs
