+++ src/render/lighting/lighting_projection.rs
