+++ src/render/lighting/mod.rs
@@ -1,5 +1,5 @@
 pub mod lighting_real;
 pub mod lighting_simple;
-pub mod lightning_norm;
+pub mod lighting_norm;
 pub mod lighting_phong;
-pub mod lightning_projection;
+pub mod lighting_projection;
